third_party/WebKit/LayoutTests/fast/html/imports/import-same-url.html

<!DOCTYPE html>
<html>
<script src="../../../resources/js-test.js"></script>
<head>
<link id="target1" rel="import" href="resources/hello.html">
<link id="target2" rel="import" href="resources/hello.html">
<link id="target3" rel="import" href="resources/hello.html#fragment">
</head>
<body>
<script>
shouldBeNonNull("target1.import");
shouldBe("target1.import", "target2.import");
shouldBe("target1.import", "target3.import");
</script>
</body>
</html>
